+++ internal/service/ref/branches_test.go
@@ -208,3 +208,99 @@ func TestFailedDeleteBranchRequest(t *testing.T) {
 		})
 	}
 }
+
+func TestSuccessfulFindBranchRequest(t *testing.T) {
+	server := runRefServiceServer(t)
+	defer server.Stop()
+
+	client, conn := newRefClient(t)
+	defer conn.Close()
+
+	branchNameInput := "master"
+	branchTarget, err := log.GetCommit(context.Background(), testRepo, branchNameInput, "")
+	require.NoError(t, err)
+
+	branch := &pb.Branch{
+		Name:         []byte(branchNameInput),
+		TargetCommit: branchTarget,
+	}
+
+	testCases := []struct {
+		desc           string
+		branchName     string
+		expectedBranch *pb.Branch
+	}{
+		{
+			desc:           "regular branch name",
+			branchName:     branchNameInput,
+			expectedBranch: branch,
+		},
+		{
+			desc:           "absolute reference path",
+			branchName:     "refs/heads/" + branchNameInput,
+			expectedBranch: branch,
+		},
+		{
+			desc:           "heads path",
+			branchName:     "heads/" + branchNameInput,
+			expectedBranch: branch,
+		},
+		{
+			desc:       "non-existent branch",
+			branchName: "i-do-not-exist-on-this-repo",
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.desc, func(t *testing.T) {
+			request := &pb.FindBranchRequest{
+				Repository: testRepo,
+				Name:       []byte(testCase.branchName),
+			}
+
+			ctx, cancel := context.WithCancel(context.Background())
+			defer cancel()
+
+			response, err := client.FindBranch(ctx, request)
+
+			require.NoError(t, err)
+			require.Equal(t, testCase.expectedBranch, response.Branch, "mismatched branches")
+		})
+	}
+}
+
+func TestFailedFindBranchRequest(t *testing.T) {
+	server := runRefServiceServer(t)
+	defer server.Stop()
+
+	client, conn := newRefClient(t)
+	defer conn.Close()
+
+	testCases := []struct {
+		desc       string
+		branchName string
+		code       codes.Code
+	}{
+		{
+			desc:       "empty branch name",
+			branchName: "",
+			code:       codes.InvalidArgument,
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.desc, func(t *testing.T) {
+
+			request := &pb.FindBranchRequest{
+				Repository: testRepo,
+				Name:       []byte(testCase.branchName),
+			}
+
+			ctx, cancel := context.WithCancel(context.Background())
+			defer cancel()
+
+			_, err := client.FindBranch(ctx, request)
+			testhelper.AssertGrpcError(t, err, testCase.code, "")
+		})
+	}
+}